match: func_8034D4AC

+++ include/uv_dobj.h
@@ -15,7 +15,7 @@ typedef struct {
 extern Unk80263780 D_80263780[8];
 
 u16 func_8021731C(void);
-void uvDobjModel(s32, u16);
+void uvDobjModel(s32, s32);
 void uvDobjPosm(u16, s32, void*);
 void uvDobjState(s32, s32);
 void uvDobjProps(s32, s32, f64, s32);
